+++ src/main/java/com/seowonn/mymap/service/Impl/VisitLogServiceImpl.java
@@ -45,16 +45,11 @@ public class VisitLogServiceImpl implements VisitLogService {
   private final VisitLogRepository visitLogRepository;
   private final MyMapRepository myMapRepository;
   private final SiGunGuRepository siGunGuRepository;
-  private final MemberRepository memberRepository;
-  private final LikesRepository likesRepository;
 
   private final S3ServiceImpl s3Service;
   private final CheckService checkService;
   private final MyMapServiceImpl myMapService;
   private final RedisServiceImpl redisService;
-
-  @PersistenceContext
-  private EntityManager entityManager;
 
   @Override
   @Transactional
@@ -176,55 +171,4 @@ public class VisitLogServiceImpl implements VisitLogService {
         .orElseThrow(() -> new MyMapSystemException(VISIT_LOG_NOT_FOUND));
   }
 
-  @Override
-  @Transactional
-  public VisitLog addVisitLogLikes(Long myMapId, Long visitLogId) {
-
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    String userId = authentication.getName();
-
-    VisitLog visitLog = visitLogRepository.findById(visitLogId)
-        .orElseThrow(() -> new MyMapSystemException(VISIT_LOG_NOT_FOUND));
-
-    Member member = memberRepository.findByUserId(userId)
-        .orElseThrow(() -> new MyMapSystemException(USER_NOT_FOUND));
-
-    // 이미 좋아요를 누른 사람인지 확인
-    Optional<Likes> likesOptional =
-        likesRepository.findByVisitLogAndVisitId(visitLog, member.getId());
-
-    if(likesOptional.isEmpty()){
-      Likes likes = Likes.of(member.getId(), visitLog);
-      likesRepository.save(likes);
-      visitLogRepository.addLikes(visitLogId);
-      entityManager.refresh(visitLog);  // VisitLog 엔티티 새로고침
-    }
-
-    return visitLog;
-  }
-
-  @Override
-  @Transactional
-  public VisitLog deleteVisitLogLikes(Long myMapId, Long visitLogId) {
-
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    String userId = authentication.getName();
-
-    Member member = memberRepository.findByUserId(userId)
-        .orElseThrow(() -> new MyMapSystemException(USER_NOT_FOUND));
-
-    VisitLog visitLog = visitLogRepository.findById(visitLogId)
-        .orElseThrow(() -> new MyMapSystemException(VISIT_LOG_NOT_FOUND));
-
-    Likes likes = likesRepository.findByVisitLogAndVisitId(visitLog,
-            member.getId())
-        .orElseThrow(() -> new MyMapSystemException(NO_LIKES_CLICKED));
-
-    likesRepository.delete(likes);
-    if(visitLog.getLikes() > 0){
-      visitLogRepository.deleteLikes(visitLogId);
-      entityManager.refresh(visitLog);  // VisitLog 엔티티 새로고침
-    }
-    return visitLog;
-  }
 }
