+++ src/main/java/com/seowonn/mymap/service/VisitorService.java
@@ -1,0 +1,11 @@
+package com.seowonn.mymap.service;
+
+import com.seowonn.mymap.entity.VisitLog;
+
+public interface VisitorService {
+
+  VisitLog addVisitLogLikes(Long myMapId, Long visitLogId);
+
+  VisitLog deleteVisitLogLikes(Long myMapId, Long visitLogId);
+
+}
